another rename cleanup
