+++ lucien/src/main/java/com/hellblazer/luciferase/lucien/cast/SphericalPublish.java
@@ -14,18 +14,18 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hellblazer.luciferase.sentinel.cast;
+package com.hellblazer.luciferase.lucien.cast;
 
 import javax.vecmath.Point3f;
 
 /**
  * @author hal.hildebrand
  */
-public class SphericalSubscription extends SpatialSubscription {
+public class SphericalPublish extends SpatialPublish {
     protected float radius;
     protected float radiusSquared;
 
-    public SphericalSubscription(Point3f location, float radius) {
+    public SphericalPublish(Point3f location, float radius) {
         super(location);
         this.radius = radius;
         this.radiusSquared = radius * radius;
